+++ lib/src/contracts/uhst_event_handlers.dart
@@ -1,7 +1,7 @@
 library uhst;
 
-import 'contracts/uhst_socket.dart';
-import 'models/message.dart';
+import '../models/message.dart';
+import 'uhst_socket.dart';
 
 typedef void OpenHandler({required String data});
 typedef void MessageHandler({required Message? message});
