+++ lib/src/contracts/uhst_socket.dart
@@ -5,7 +5,7 @@ import 'dart:typed_data';
 import 'package:universal_html/html.dart';
 
 import '../models/message.dart';
-import '../uhst_event_handlers.dart';
+import 'uhst_event_handlers.dart';
 
 abstract class UhstSocket {
   void onOpen({required OpenHandler handler});
