+++ lib/src/utils/uhst_errors.dart
