+++ lib/src/utils/uhst_exceptions.dart
