+++ lib/uhst.dart
@@ -6,10 +6,10 @@ library uhst;
 
 export 'src/contracts/uhst_host_socket.dart';
 export 'src/contracts/uhst_socket.dart';
+export 'src/hosts/uhst_host.dart';
 export 'src/models/message.dart';
-export 'src/relay_socket_provider.dart';
+export 'src/sockets/relay_socket_provider.dart';
+export 'src/sockets/web_rtc_socket_provider.dart';
 export 'src/uhst.dart';
-export 'src/uhst_errors.dart';
-export 'src/uhst_exceptions.dart';
-export 'src/uhst_host.dart';
-export 'src/web_rtc_socket_provider.dart';
+export 'src/utils/uhst_errors.dart';
+export 'src/utils/uhst_exceptions.dart';
