+++ test/src/relay_socket_provider_test.dart
@@ -1,7 +1,7 @@
 import 'package:test/test.dart';
-import 'package:uhst/src/api_client.dart';
+import 'package:uhst/src/api_clients/api_client.dart';
 import 'package:uhst/src/models/socket_params.dart';
-import 'package:uhst/src/relay_socket_provider.dart';
+import 'package:uhst/uhst.dart';
 
 void main() {
   group('# RelaySocketProvider', () {
